media: venus: fix possible buffer overlow casued bad DMA value in venus_sfr_print()
The value hdev->sfr.kva is stored in DMA memory, and it is assigned to
sfr, so sfr->buf_size can be modified at anytime by malicious hardware.
In this case, a buffer overflow may happen when the code
"sfr->data[sfr->buf_size - 1]" is executed.

To fix this possible bug, sfr->buf_size is assigned to a local variable,
and then this variable is checked before being used.

Signed-off-by: Jia-Ju Bai <baijiaju1990@gmail.com>