ath10k: Pause the tx queues when firmware is down
When the FW goes down, the transmission of
packets cannot be done. Currently we are not
pausing the tx queues, hence there will be too
many attempts to trasmit packets and each of
them resulting in failure and thereby logging
unwanted failure messages onto the console,
as shown below

[   60.842192] ath10k_snoc 18800000.wifi: failed to transmit packet, dropping: -108
[   60.850181] ath10k_snoc 18800000.wifi: failed to submit frame: -108
[   60.857058] ath10k_snoc 18800000.wifi: failed to push frame: -108
[   60.936217] ath10k_snoc 18800000.wifi: failed to transmit packet, dropping: -108
[   60.944496] ath10k_snoc 18800000.wifi: failed to submit frame: -108
[   60.951265] ath10k_snoc 18800000.wifi: failed to push frame: -108
[   61.976589] ath10k_snoc 18800000.wifi: failed to transmit packet, dropping: -108

Stop the tx queues when the wlan firmware is
down and also remove the unwanted duplicate/extra
logging in the packet transmission path.

Tested-on: WCN3990 hw1.0 SNOC WLAN.HL.3.1-01040-QCAHLSWMTPLZ-1

Signed-off-by: Rakesh Pillai <pillair@codeaurora.org>
Signed-off-by: Steev Klimaszewski <steev@kali.org>