iommu: arm-smmu: qcom: Don't use identity domain for display
As MDSS probes it will be added to an iommu domain of type IDENTITY and
BYPASS streams with be setup, but as the hypervisor kindly changes this
to a FAULT the next screen update causes the device to crash.

Instead just return 0, which will result in a normal domain - with
translation - being setup. This results in a translation fault, which is
reported in the log, but allows MDSS to continue.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>